// server/src/enums/permission.enum.ts
export enum PermissionEnum {
  VIEW_SELF = 'VIEW_SELF',
  VIEW_USER = 'VIEW_USER',
  VIEW_USERS = 'VIEW_USERS',
  VIEW_PERMISSIONS = 'VIEW_PERMISSIONS',
  EDIT_SELF = 'EDIT_SELF',
  EDIT_USER = 'EDIT_USER',
  EDIT_PERMISSIONS = 'EDIT_PERMISSIONS',
  DELETE_SELF = 'DELETE_SELF',
  DELETE_USER = 'DELETE_USER',
  DELETE_PERMISSIONS = 'DELETE_PERMISSIONS',
}